CRI: eight more defects closed off the named worklist
Working build/adx_named.py's list, where the address is already settled by
a shipped symbol table so a mismatch is a defect rather than a mispairing.
Named-exact 84 -> 92, right-size-and-differing 50 -> 42.

  ADXPD_OBJ carries 0x6C more before its extended parameters:
  ADXPD_SetExtPrm stores them at 0xA0/0xA2/0xA4 where this layout has
  0x34/0x36/0x38. `dly` at 0x28 is NOT affected, which is why the padding
  goes after it - ADXPD_GetDly and ADXPD_SetDly already matched byte for
  byte, and that is what fixes the insertion point rather than guessing.

  DTR_OBJ 0x40 -> 0x80, from DTR_Init's `addiu $a2, $zero, 0x800`.
  DTX scans a pool of 16, not 8 (`addiu $s1, $zero, 0xf`, two loops).
  sjx_sndcbf scans 32 clients, not 16 (`slti $v0, $s3, 0x20`).
  LSC reports its parameter error as -3, not -1, in both trees - 15 sites.
  ADX_DecodeInfoSpsd reads the channel count SIGNED: retail `lb`, and the
  `(Uint8)` cast in the reconstruction forced `lbu`.
  MWSFSVM_GotoIdleBorder asks for border 7.

Rejected after measuring: -mno-branch-likely. Several functions differ
only in `bnel` where retail has `bne`, but the CRI region contains 1512
branch-likely instructions, so the flag would break far more than it
fixed. Those are loop shape and have to go one at a time.

168 markers, 168 MATCH. Tree-wide 8247 -> 8248 of 13102.

The gap between "168 markers" and "92 named-exact" is worth stating: most
markers sit at addresses no shipped symbol table names, and most named
functions are in the 409 that are still a different size. Neither number
contains the other.

Image 3d1d3d2b, SLUS 4eeec036, 706 tests, reconcile clean.

+++ src/cri/libadxe/adx_bsps.c
@@ -33,7 +33,7 @@ Sint32 ADX_DecodeInfoSpsd(Sint8 *ibuf, Sint32 ibuflen, Sint16 *dlen, Sint8 *code
     case 1:
         *bps = 8;
         
-        *blksize = temp = (Uint8)*nch;
+        *blksize = temp = *nch;
         
         *nsmpl_blk = 1;
         
@@ -45,7 +45,7 @@ Sint32 ADX_DecodeInfoSpsd(Sint8 *ibuf, Sint32 ibuflen, Sint16 *dlen, Sint8 *code
     case 3:
         *bps = 4;
         
-        *blksize = temp2 = (Uint8)*nch;
+        *blksize = temp2 = *nch;
         
         *nsmpl_blk = 2;
